fix some react doctor things
