+++ src/renderer/components/Experiment/Tasks/QueueTaskModal/index.tsx
@@ -25,7 +25,8 @@ import { useAuth } from 'renderer/lib/authContext';
 import SweepConfigSection from '../SweepConfigSection';
 import SafeJSONParse from 'renderer/components/Shared/SafeJSONParse';
 import { getPreferredProviderId } from '../providerCompatibility';
-import SlurmFlagsSection, { createSlurmFlag } from './SlurmFlagsSection';
+import SlurmFlagsSection from './SlurmFlagsSection';
+import { createSlurmFlag } from './slurmFlags';
 import TrackingSection from './TrackingSection';
 import ProfilingSection from './ProfilingSection';
 import DescriptionSection from './DescriptionSection';
